fix: remove GCF transaction fallback
I would like to remove the code to run a transactional commit on GCF periodically. I hope this code is not needed anymore since the socket timeout has been drastically increased since our beta release and GCF should no longer block our outgoing traffic. On top of that, we never used a transaction runner for these commits, so retries with code ABORTED were not possible. This code was also not added to BulkWriter.